adding some logic to retrieve and delete the transaction record base on user's login

+++ public/handler/transaction_payment_process.php
@@ -8,12 +8,24 @@
 </head>
 <body>
 <?php 
+
 include '../Database/dbconfig.php';
-    
+
+session_start();
+
+//CHECK IS USER IS LOGGED IN
+if(!isset($_SESSION['user_id'])) {
+  header('Location: ../Logout/home_page.php');
+  exit();
+}
+
+$userId = $_SESSION['user_id'];
 $rows = [];
 
-    $transactionData = "SELECT * FROM process_buyticket";
+
+    $transactionData = "SELECT * FROM process_buyticket WHERE user_id = ?";
     $stmt = $connection->prepare($transactionData);
+    $stmt -> bind_param('i', $userId);
     $stmt->execute();
     $result = $stmt->get_result();
     
@@ -24,9 +36,9 @@ $rows = [];
         //DELETE TRANSACTION RECORD BASE ON ID
         if(isset($_POST['delete'])) {
             $id_to_delete = $_POST['id_to_delete'];
-            $deleteID = "DELETE FROM process_buyticket WHERE id = ?";
+            $deleteID = "DELETE FROM process_buyticket WHERE id = ? AND user_id = ?";
             $stmt = $connection->prepare($deleteID);
-            $stmt->bind_param('i', $id_to_delete);
+            $stmt->bind_param('ii', $id_to_delete,$userId);
             $stmt->execute();
             
             if($stmt->execute()) {
